refac: mv checkpoint loading logic to new module
