+++ src/careamics/lightning/dataset_ng/lightning_modules/get_module.py
@@ -1,9 +1,5 @@
 """Factory functions for lightning modules."""
 
-from pathlib import Path
-
-import torch
-
 from careamics.config import CAREAlgorithm, N2VAlgorithm
 from careamics.config.algorithms.unet_algorithm_config import UNetBasedAlgorithm
 from careamics.config.support import SupportedAlgorithm
@@ -74,39 +70,3 @@ def get_module_cls(algorithm: SupportedAlgorithm) -> CAREamicsModuleCls:
             raise NotImplementedError(
                 f"Support for {algorithm.value} has not been implemented yet."
             )
-
-
-def load_module_from_checkpoint(checkpoint_path: Path) -> CAREamicsModule:
-    """
-    Load a trained CAREamics module from checkpoint.
-
-    Automatically detects the algorithm type from the checkpoint and loads
-    the appropriate module with trained weights.
-
-    Parameters
-    ----------
-    checkpoint_path : Path
-        Path to the PyTorch Lightning checkpoint file.
-
-    Returns
-    -------
-    CAREamicsModule
-        Lightning module with loaded weights.
-
-    Raises
-    ------
-    ValueError
-        If the algorithm type cannot be determined from the checkpoint.
-    """
-    checkpoint = torch.load(checkpoint_path, map_location="cpu")
-
-    try:
-        algorithm = checkpoint["hyper_parameters"]["algorithm_config"]["algorithm"]
-        algorithm = SupportedAlgorithm(algorithm)
-    except (KeyError, ValueError) as e:
-        raise ValueError(
-            f"Could not determine algorithm type from checkpoint at: {checkpoint_path}"
-        ) from e
-
-    ModuleClass = get_module_cls(algorithm)
-    return ModuleClass.load_from_checkpoint(checkpoint_path)
